refactor: move interpolate_grid and convert_to_2d_grid from KrigingModel to spatial_utils.py
